FIRE-Agent/requirements.txt
# Core Dependencies
google-genai
python-dotenv>=1.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# Utilities
pydantic-settings>=2.0.0
requests>=2.31.0
